--- SHEET_SCHEMA.yaml
@@ -1,10 +1,0 @@
-leadSheetColumns:
-    companyName: "company_name"
-    companyNumber: "company_number"
-    address: "address_string"
-    dateCreated: "date_of_creation"
-    companyType: "company_type"
-    sicCodes: "sic_codes_string"
-    runId: "added_on_run_id"
-    runTs: "added_run_ts"
-    officerNames: "company_officer_names"
