--- constants.py
@@ -1,10 +1,0 @@
-SEARCH_URL = "https://api.company-information.service.gov.uk/advanced-search/companies"
-LIMIT = 600 #move to config? #600 actual value: https://developer-specs.company-information.service.gov.uk/guides/rateLimiting
-RATE = 300 #move to config? #300 check above
-LOG_FORMAT = "[%(asctime)s - %(levelname)s - %(funcName)s()]\t%(message)s" #TO-DO: show function / more context here
-LOG_FILE_NAME = "leadTracker.log"
-LOG_FOLDER = "Logs"
-TIMEZONE = "UTC"
-SHEET_SCHEMA_PATH = "SHEET_SCHEMA.yaml"
-CACHE_DB = "cache.db"
-OFFICER_URL = ""
