--- environment.yaml
@@ -1,7 +1,0 @@
-#TO-DO: Migrate ENV variables from .env here
-#TO-DO: make sure this is used in validation steps
-
-request_types:
-  search: "search"
-  officers: "officers"
-  company: "company"
